Move files around and update references
