--- settings/__init__.py
@@ -1,1 +1,0 @@
-# PLACEHOLDER, DO NOT DELETE THIS FILE
